Extract Flipgrid source
